+++ data/src/main/java/camp/nextstep/edu/github/data/GithubRepositoryImpl.kt
@@ -2,20 +2,18 @@ package camp.nextstep.edu.github.data
 
 import camp.nextstep.edu.github.domain.GithubRepository
 import camp.nextstep.edu.github.domain.GithubService
+import camp.nextstep.edu.github.domain.Repo
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
-internal class GithubRepositoryImpl: GithubRepository {
-    companion object {
-        private const val BASE_URL = "https://api.github.com"
-    }
+internal class GithubRepositoryImpl(
+    private val gitHubService: GithubService
+) : GithubRepository {
+//    fun getGithubService(): GithubService {
+//        return retrofit().create(GithubService::class.java)
+//    }
 
-    private fun retrofit(): Retrofit = Retrofit.Builder()
-        .baseUrl(BASE_URL)
-        .addConverterFactory(GsonConverterFactory.create())
-        .build()
-
-    override suspend fun getGithubService(): GithubService {
-        return retrofit().create(GithubService::class.java)
+    override suspend fun getRepositories(): List<Repo> {
+        TODO("Not yet implemented")
     }
 }
