+++ domain/src/main/java/camp/nextstep/edu/github/domain/GithubRepository.kt
@@ -1,5 +1,6 @@
 package camp.nextstep.edu.github.domain
 
+
 interface GithubRepository {
-    suspend fun getGithubService(): GithubService
+    suspend fun getRepositories(): List<Repo>
 }
