+++ domain/src/main/java/camp/nextstep/edu/github/domain/GithubService.kt
@@ -1,8 +1,9 @@
 package camp.nextstep.edu.github.domain
 
+import retrofit2.Call
 import retrofit2.http.GET
 
 interface GithubService {
     @GET("/repositories")
-    suspend fun listRepos(): List<Repo>
+    fun listRepos(): Call<List<Repo>>
 }
